Sphere with Bresenham's algorithm done.

+++ src/utils/ParticleGenerator.cpp
@@ -152,7 +152,7 @@ void ParticleGenerator::extractSpheres(const string filename){
 			utils::Vector<double, 3> centerP(a);
 			double b[] = {i->startVel().vX(), i->startVel().vY(), i->startVel().vZ()};
 			utils::Vector<double, 3> vel(b);
-			int radiuss = (int) i->radiussph();
+			int radiuss = i->radiussph();
 		
 	      		Sphere s(centerP, vel, meanV, m, radiuss, mesh);
 			sphereList.push_back(s);
